d.ts error fixed
