+++ src/apiServices/auth/route.ts
@@ -2,7 +2,7 @@ import express from "express";
 import checkAuth from "../../middlewares/checkAuth";
 import checkUserType from "../../middlewares/checkUserType";
 import checkRoleAuth from "../../middlewares/roleAuth";
-import { Role, Type } from "../../types/data";
+import { Role, Type } from "../../types/data/enums";
 import { loginAdmin, logoutProveedor, registerProveedor, confirmProveedorAccount, registerUsuario, loginProveedor, logoutUsuario, confirmAccount, loginUsuario } from "./controller";
 import { validateUserLogin, validateUserRegister, validateCode } from "../../middlewares/validator";
 const router=express.Router();
