+++ src/apiServices/licitacion/route.ts
@@ -3,7 +3,7 @@ import checkAuth from "../../middlewares/checkAuth";
 import checkUserType from "../../middlewares/checkUserType";
 //import { upload } from "../../middlewares/filesLicitacion/uploadFile";
 import checkRoleAuth from "../../middlewares/roleAuth";
-import { Role, Type } from "../../types/data";
+import { Role, Type } from "../../types/data/enums";
 import { licitacionId, showLicitacionById, showLicitaciones, showLicitacionesFree, getTipos, createLicitacion, updateLicitacion } from "./controller";
 const router=express.Router();
