+++ src/apiServices/servicio/route.ts
@@ -2,7 +2,7 @@ import express from "express";
 import checkAuth from "../../middlewares/checkAuth";
 import checkUserType from "../../middlewares/checkUserType";
 import checkRoleAuth from "../../middlewares/roleAuth";
-import { Role, Type } from "../../types/data";
+import { Role, Type } from "../../types/data/enums";
 const router=express.Router();
 import { getServicios, addServicio } from "./controller";
 router.get("/getServicios",checkAuth,checkUserType([Type.User]),checkRoleAuth([Role.Basico,Role.Premium,Role.Admin]),getServicios);
