+++ src/types/data/enums.ts
@@ -1,0 +1,14 @@
+export enum Role{
+    Basico="basico",
+    Premium="premium",
+    Admin="admin"
+}
+export enum Type{
+    User="user",
+    Proveedor="proveedor"
+}
+export enum Estado{
+    Online="online",
+    Offline="offline",
+    ToConfirm="toConfirm"
+}
