+++ src/types/form/enums.ts
@@ -1,0 +1,4 @@
+export enum Estado{
+    Abierto="abierto",
+    Cerrado="cerrado"
+}
