+++ src/types/methods/index.d.ts
@@ -1,5 +1,6 @@
 import { RequestHandler } from "express";
-import { DataProveedorToken, DataUserToken, Proveedor, Role, Type, User } from "../data";
+import { DataProveedorToken, DataUserToken, Proveedor, User } from "../data";
+import { Role, Type } from "../data/enums";
 
 export type SignToken=(user:User|Proveedor)=>string
 export type VerifyToken=(token:string)=>null|DataUserToken|DataProveedorToken
